+++ src/k8s_sandbox/__init__.py
@@ -3,6 +3,7 @@
 from k8s_sandbox._error import K8sError
 from k8s_sandbox._pod import GetReturncodeError, PodError
 from k8s_sandbox._pod.error import ContainerRestartedError, PodReplacedError
+from k8s_sandbox._priority import PrioritySourceJob
 from k8s_sandbox._sandbox_environment import (
     K8sSandboxEnvironment,
     K8sSandboxEnvironmentConfig,
@@ -16,4 +17,5 @@ __all__ = [
     "K8sSandboxEnvironmentConfig",
     "PodError",
     "PodReplacedError",
+    "PrioritySourceJob",
 ]